// src/app/contentManager/enumerations/mediaInfo.enum.ts
export enum MediaInfoField {
    VIDEO_LINK = 'video_link',
    VIDEO_ID = 'video_id',
    TITLE = 'title',
    DATE = 'date',
    DIRECTOR = 'director',
    DIRECTOR_LINK = 'director_link',
    MAKER = 'maker',
    MAKER_LINK = 'maker_link',
    LABEL = 'label',
    LABEL_LINK = 'label_link',
    VIDEO_GENRES = 'video_genres',
    VIDEO_CASTS = 'video_casts',
    CAST_NUMBER = 'cast_number',
    IS_OMNIBUS = 'is_omnibus',
    HTML = 'html',
    LENGTH = 'length',
    IMG_URL = 'img_url',
    GOT_IT = 'got_it',
    UPDATE_TIME = 'update_time',
    VIDEO_SCORE = 'video_score',
    VIDEO_PATH = 'video_path',
}